# RedoclyDemo/OpenAPI.yaml
openapi: 3.0.1
info:
  title: RedoclyDemo API
  version: v1
paths:
  /Calculator/add:
    get:
      tags:
        - Calculator
      parameters:
        - name: left
          in: query
          schema:
            type: number
            format: double
        - name: right
          in: query
          schema:
            type: number
            format: double
      responses:
        '200':
          description: OK
          content:
            text/plain:
              schema:
                type: number
                format: double
            application/json:
              schema:
                type: number
                format: double
            text/json:
              schema:
                type: number
                format: double
  /Calculator/subtract:
    get:
      tags:
        - Calculator
      parameters:
        - name: left
          in: query
          schema:
            type: number
            format: double
        - name: right
          in: query
          schema:
            type: number
            format: double
      responses:
        '200':
          description: OK
          content:
            text/plain:
              schema:
                type: number
                format: double
            application/json:
              schema:
                type: number
                format: double
            text/json:
              schema:
                type: number
                format: double
  /Calculator/multiply:
    get:
      tags:
        - Calculator
      parameters:
        - name: left
          in: query
          schema:
            type: number
            format: double
        - name: right
          in: query
          schema:
            type: number
            format: double
      responses:
        '200':
          description: OK
          content:
            text/plain:
              schema:
                type: number
                format: double
            application/json:
              schema:
                type: number
                format: double
            text/json:
              schema:
                type: number
                format: double
  /Calculator/divide:
    get:
      tags:
        - Calculator
      parameters:
        - name: left
          in: query
          schema:
            type: number
            format: double
        - name: right
          in: query
          schema:
            type: number
            format: double
      responses:
        '200':
          description: OK
          content:
            text/plain:
              schema:
                type: number
                format: double
            application/json:
              schema:
                type: number
                format: double
            text/json:
              schema:
                type: number
                format: double
  /Calculator/multipleAdd:
    get:
      tags:
        - Calculator
      parameters:
        - name: left
          in: query
          schema:
            type: number
            format: double
        - name: right
          in: query
          schema:
            type: number
            format: double
      responses:
        '200':
          description: OK
          content:
            text/plain:
              schema:
                type: number
                format: double
            application/json:
              schema:
                type: number
                format: double
            text/json:
              schema:
                type: number
                format: double
  /Calculator/addSubtractAdd:
    get:
      tags:
        - Calculator
      parameters:
        - name: left
          in: query
          schema:
            type: number
            format: double
        - name: right
          in: query
          schema:
            type: number
            format: double
      responses:
        '200':
          description: OK
          content:
            text/plain:
              schema:
                type: number
                format: double
            application/json:
              schema:
                type: number
                format: double
            text/json:
              schema:
                type: number
                format: double
  /WeatherForecast:
    get:
      tags:
        - WeatherForecast
      operationId: GetWeatherForecast
      responses:
        '200':
          description: OK
          content:
            text/plain:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/WeatherForecast'
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/WeatherForecast'
            text/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/WeatherForecast'
components:
  schemas:
    WeatherForecast:
      type: object
      properties:
        date:
          type: string
          format: date
        temperatureC:
          type: integer
          format: int32
        temperatureF:
          type: integer
          format: int32
          readOnly: true
        summary:
          type: string
          nullable: true
      additionalProperties: false